requirements.txt
gunicorn==22.0.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
Flask==3.1.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
urllib3==2.5.0
slack-sdk==3.33.4
Werkzeug==3.1.3
